+++ CHANGELOG.md
@@ -103,7 +103,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - The deprecated `SemVersion` function is removed in `go.opentelemetry.io/contrib/instrumentation/host` package, use `Version` instead. (#7203)
 - The `GRPCStatusCodeKey` constant from `go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc` is removed.
   Use `semconv.RPCGRPCStatusCodeKey` from `go.opentelemetry.io/otel/semconv/*` instead. (#7270)
-- In `go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin`, mark the `Filter` and `MetricAttributeFn` types as well as the `WithFilter` and `WithMetricAttributeFn` functions as deprecated. (#7244)
+
+### Deprecated
+
+- Deprecate `WithFilter` option and `Filter` type in `go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin`. Use `GinFilter` option instead. (#7244)
+- Deprecate `WithMetricAttributeFn` option and `MetricAttributeFn` type in `go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin`. Use `WithGinMetricAttributeFn` option instead. (#7244)
 
 <!-- Released section -->
 <!-- Don't change this section unless doing release -->